Rename followers python package
This allows to call from other packages e.g. `from followers.pid import PID`
